MER-5795 Extract experiment context services
